Add analysis viewer action types and creators
