+++ client/src/js/analyses/components/Pathoscope/Viewer.js
@@ -1,94 +1,191 @@
-import {compact, flatMap, map, max, maxBy, mean, round, sortBy, sum, sumBy} from "lodash-es";
-import PropTypes from "prop-types";
 import React from "react";
+import PropTypes from "prop-types";
+import { filter, map, sortBy, sum, xor } from "lodash-es";
+import { Icon, Flex, FlexItem, Button, Checkbox } from "../../../base/index";
+import { Row, Col, Dropdown, MenuItem, FormGroup, InputGroup, FormControl } from "react-bootstrap";
 
-import {Alert} from "../../../base/index";
-import {formatIsolateName} from "../../../utils";
-import {fillAlign, median} from "../../utils";
-import PathoscopeController from "./Controller";
+import PathoscopeList from "./List";
 
-const PathoscopeViewer = (props) => {
+export default class PathoscopeController extends React.Component {
 
-    if (props.diagnosis.length > 0) {
+    filter = (eventKey) => {
 
-        const mappedReadCount = props.read_count;
+        if (eventKey === "OTUs") {
+            return this.setState({filterOTUs: !this.state.filterOTUs});
+        }
 
-        const data = map(props.diagnosis, baseOTU => {
-            // Go through each isolate associated with the OTU, adding properties for weight, read count,
-            // mean depth, and coverage. These values will be calculated from the sequences owned by each isolate.
-            let isolates = map(baseOTU.isolates, isolate => {
-                // Make a name for the isolate by joining the source type and name, eg. "Isolate" + "Q47".
-                let name = formatIsolateName(isolate);
+        if (eventKey === "isolates") {
+            return this.setState({filterIsolates: !this.state.filterIsolates});
+        }
 
-                if (name === "unknown unknown") {
-                    name = "Unnamed Isolate";
-                }
+        const filterValue = !(this.state.filterOTUs && this.state.filterIsolates);
 
-                const sequences = map(isolate.sequences, sequence => {
-                    const filled = fillAlign(sequence);
-
-                    return {
-                        ...sequence,
-                        reads: round(sequence.pi * mappedReadCount),
-                        meanDepth: mean(filled),
-                        medianDepth: median(filled),
-                        sumDepth: sum(filled),
-                        filled
-                    };
-                });
-
-                const filled = flatMap(sequences, "filled");
-
-                const length = filled.length;
-
-                const coverage = compact(filled).length / length;
-
-                return {
-                    ...isolate,
-                    name,
-                    length,
-                    pi: sumBy(sequences, "pi"),
-                    reads: sumBy(sequences, "reads"),
-                    maxDepth: max(filled),
-                    meanDepth: mean(filled),
-                    medianDepth: median(filled),
-                    coverage: coverage.toFixed(3)
-                };
-            });
-
-            isolates = sortBy(isolates, "coverage").reverse();
-
-            const pi = sumBy(isolates, "pi");
-
-            return {
-                ...baseOTU,
-                coverage: maxBy(isolates, "coverage").coverage,
-                isolates,
-                maxGenomeLength: maxBy(isolates, "length").length,
-                maxDepth: maxBy(isolates, "meanDepth").maxDepth,
-                meanDepth: maxBy(isolates, "meanDepth").meanDepth,
-                medianDepth: maxBy(isolates, "medianDepth").medianDepth,
-                pi,
-                reads: pi * mappedReadCount
-            };
+        return this.setState({
+            filterOTUs: filterValue,
+            filterIsolates: filterValue
         });
+    };
 
-        return <PathoscopeController data={data} maxReadLength={props.maxReadLength} />;
+    render () {
+
+        let data = sortBy(this.props.data, this.state.sortKey);
+
+        const re = this.state.findTerm ? new RegExp(this.state.findTerm, "i") : null;
+
+        if (this.state.filterOTUs) {
+            const totalReadsMapped = sum(map(data, "reads"));
+
+            data = filter(data, otu => (
+                (otu.pi * totalReadsMapped >= otu.length * 0.8 / this.props.maxReadLength) &&
+                (!re || (re.test(otu.abbreviation) || re.test(otu.name)))
+            ));
+        } else {
+            data = filter(data, (otu) => !re || (re.test(otu.abbreviation) || re.test(otu.name)));
+        }
+
+        if (this.state.filterIsolates) {
+            data = map(data, otu => ({
+                ...otu,
+                isolates: filter(otu.isolates, isolate => (isolate.pi >= 0.03 * otu.pi))
+            }));
+        }
+
+        if (this.state.sortDescending) {
+            data.reverse();
+        }
+
+        return (
+            <div>
+                <Row>
+                    <Col xs={12} md={7}>
+                        <FormGroup>
+                            <InputGroup>
+                                <InputGroup.Addon>
+                                    <Icon name="search" />
+                                </InputGroup.Addon>
+                                <FormControl
+                                    value={this.state.findTerm}
+                                    onChange={(e) => this.setState({findTerm: e.target.value})}
+                                />
+                            </InputGroup>
+                        </FormGroup>
+                    </Col>
+
+                    <Col xs={12} md={5}>
+                        <div className="toolbar">
+                            <FormGroup>
+                                <InputGroup>
+                                    <InputGroup.Button>
+                                        <Button
+                                            title="Sort Direction"
+                                            onClick={this.toggleSortDescending}
+                                            tip="Sort List"
+                                        >
+                                            <Icon
+                                                name={this.state.sortDescending ? "sort-amount-down" : "sort-amount-up"}
+                                            />
+                                        </Button>
+                                    </InputGroup.Button>
+                                    <FormControl
+                                        componentClass="select"
+                                        value={this.state.sortKey}
+                                        onChange={this.setSortKey}
+                                    >
+                                        <option className="text-primary" value="coverage">Coverage</option>
+                                        <option className="text-success" value="pi">Weight</option>
+                                        <option className="text-danger" value="maxDepth">Depth</option>
+                                    </FormControl>
+                                </InputGroup>
+                            </FormGroup>
+
+                            <Button
+                                icon="compress"
+                                title="Collapse"
+                                tip="Collapse Opened"
+                                onClick={this.collapseAll}
+                                className="hidden-xs"
+                                disabled={this.state.expanded.length === 0}
+                            />
+
+                            <Button
+                                icon="chart-pie"
+                                title="Change Weight Format"
+                                tip="Change Weight Format"
+                                active={!this.state.showReads}
+                                className="hidden-xs"
+                                onClick={this.toggleShowReads}
+                            />
+
+                            <Button
+                                icon="chart-bar"
+                                title="Show Isolate Median"
+                                tip="Show Isolate Median"
+                                active={this.state.showMedian}
+                                className="hidden-xs"
+                                onClick={this.toggleShowMedian}
+                            />
+
+                            <Button
+                                icon="crop"
+                                title="Crop Outliers"
+                                tip="Crop Outliers"
+                                active={this.state.isCrop}
+                                className="hidden-xs"
+                                onClick={this.toggleOutlier}
+                            />
+
+                            <Dropdown
+                                id="job-clear-dropdown"
+                                onSelect={this.handleSelect}
+                                className="split-dropdown"
+                                pullRight
+                                style={{zIndex: "1"}}
+                            >
+                                <Button
+                                    title="Filter"
+                                    tip="Filter Results"
+                                    onClick={this.filter}
+                                    active={this.state.filterOTUs || this.state.filterIsolates}
+                                >
+                                    <Icon name="filter" />
+                                </Button>
+                                <Dropdown.Toggle />
+                                <Dropdown.Menu onSelect={this.filter}>
+                                    <MenuItem eventKey="OTUs">
+                                        <Flex>
+                                            <FlexItem>
+                                                <Checkbox checked={this.state.filterOTUs} />
+                                            </FlexItem>
+                                            <FlexItem pad={5}>
+                                                OTUs
+                                            </FlexItem>
+                                        </Flex>
+                                    </MenuItem>
+                                    <MenuItem eventKey="isolates">
+                                        <Flex>
+                                            <FlexItem>
+                                                <Checkbox checked={this.state.filterIsolates} />
+                                            </FlexItem>
+                                            <FlexItem pad={5}>
+                                                Isolates
+                                            </FlexItem>
+                                        </Flex>
+                                    </MenuItem>
+                                </Dropdown.Menu>
+                            </Dropdown>
+                        </div>
+                    </Col>
+                </Row>
+
+                <PathoscopeList
+                    data={data}
+                    expanded={this.state.expanded}
+                    toggleIn={this.toggleIn}
+                    showReads={this.state.showReads}
+                    showMedian={this.state.showMedian}
+                    isCrop={this.state.isCrop}
+                />
+            </div>
+        );
     }
-
-    return (
-        <Alert bsStyle="info" className="text-center" icon="info-circle">
-            No OTU sequences found in sample
-        </Alert>
-    );
-
-};
-
-PathoscopeViewer.propTypes = {
-    diagnosis: PropTypes.arrayOf(PropTypes.object),
-    showListing: PropTypes.bool,
-    read_count: PropTypes.number,
-    maxReadLength: PropTypes.number
-};
-
-export default PathoscopeViewer;
+}
